Refactored badge widget
