+++ rousseau_vote/lib/src/widgets/user/badge_slider.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:rousseau_vote/src/l10n/rousseau_localizations.dart';
 import 'package:rousseau_vote/src/models/profile/badge.dart';
 import 'package:rousseau_vote/src/util/profile_util.dart';
+import 'package:rousseau_vote/src/widgets/user/badge_widget.dart';
 
 class BadgeSlider extends StatelessWidget {
   const BadgeSlider({this.badges});
@@ -49,48 +50,12 @@ class BadgeSlider extends StatelessWidget {
               padding: const EdgeInsets.only(bottom: 10, left:10),
               itemCount: badgeImagesPaths.length,
               itemBuilder: (BuildContext context, int index) {
-                return Padding(
-                  padding: const EdgeInsets.symmetric(horizontal: 4, vertical: 0),
-                  child: InkWell(
-                    borderRadius: BorderRadius.circular(ICON_SIZE / 2),
-                    onTap: () => showMeritMeaning(badgeImagesPaths[index],context),
-                    child: Image( 
-                      image: AssetImage(badgeImagesPaths[index]),
-                      width: ICON_SIZE,
-                      height: ICON_SIZE,
-                    ),
-                  ),
-                );
+                return BadgeWidget(badgeImagePath: badgeImagesPaths[index]);
               }
             ),
           ),
         ]
       ),
-    );
-  }
-
-  void showMeritMeaning(String meritImagePath, BuildContext context){
-    int index = meritImagePath.indexOf('merit');
-    String meritText = meritImagePath.substring(index,index+6);
-    showDialog<CupertinoAlertDialog>(
-      context: context,
-      builder: (BuildContext context) {
-        return CupertinoAlertDialog(
-          content: Column(
-            children: <Widget>[
-              Image( 
-                image: AssetImage(meritImagePath),
-                width: 60,
-                height: 60,
-              ),
-              Padding(
-                padding: const EdgeInsets.only(top:15),
-                child: Text(RousseauLocalizations.getText(context, meritText)),
-              )
-            ],
-          ),
-        );
-      }
     );
   }
 }
